+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next Version
 ### Bug fixes
 * Fix crash when an EditorOnly tagged game object was in the exclusions list.
+* Fix non Texture2D textures that are the same reporting that they can't be combined.
 
 ## v1.4
 ### Features