+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bug fixes
 * Fix crash when an EditorOnly tagged game object was in the exclusions list.
 * Fix non Texture2D textures that are the same reporting that they can't be combined.
+* Fix non Texture2D textures trying to combine with `null`.
 
 ## v1.4
 ### Features