+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix nested BlendTrees not getting iterated through to fix animation paths.
 * Move the optimized copy to the same scene as the original.
 * Replace VFACE semantic with SV_IsFrontFace.
+* Force include TEXCOORD0 semantic in the vertex shader input struct. This fixes cases where the sematic only exists in some shader variants.
 
 ## v1.4
 ### Features