+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Move the optimized copy to the same scene as the original.
 * Replace VFACE semantic with SV_IsFrontFace.
 * Force include TEXCOORD0 semantic in the vertex shader input struct. This fixes cases where the sematic only exists in some shader variants.
+* Don't delete transforms that are moved by animations.
 
 ## v1.4
 ### Features