+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next Version
+## v1.5
 ### Bug fixes
 * Fix crash when an EditorOnly tagged game object was in the exclusions list.
 * Fix non Texture2D textures that are the same reporting that they can't be combined.